requirements.txt
requests==2.32.3
googlesearch-python==1.3.0
feedparser==6.0.11
beautifulsoup4==4.13.3
PyPDF2==3.0.1
mimetypes-magic==0.4.30